requirements.txt
streamlit==1.52.2
pillow==12.0.0
piexif==1.1.3
numpy==2.2.6
reportlab==4.4.7
